+++ packages/components-js/tests/vrt/specs/checkbox-wrapper.vrt.ts
@@ -23,7 +23,7 @@ it('should have no visual regression for :hover + :focus-visible', async () => {
     await vrt.test('checkbox-wrapper-states', async () => {
       const page = vrt.getPage();
 
-      const head = `<style type="text/css">p-checkbox-wrapper:not(:last-child) { margin-bottom: 1rem; }</style>`;
+      const head = `<style>p-checkbox-wrapper:not(:last-child) { margin-bottom: 1rem; }</style>`;
 
       const getElementsMarkup: GetMarkup = () => `
         <p-checkbox-wrapper label="Some label">