+++ packages/components/src/components/action/button-pure/button-pure.tsx
@@ -12,6 +12,7 @@ import {
   transitionListener,
   attachComponentCss,
   parseAndGetAriaAttributes,
+  isLightElectric,
 } from '../../../utils';
 import type {
   SelectedAriaAttributes,
@@ -21,7 +22,7 @@ import type {
   LinkButtonPureIconName,
   TextSize,
   TextWeight,
-  Theme,
+  ThemeExtendedElectric,
 } from '../../../types';
 import { isSizeInherit } from '../../basic/typography/text/text-utils';
 import { warnIfIsLoadingAndIconIsNone } from './button-pure-utils';
@@ -74,7 +75,7 @@ export class ButtonPure {
   @Prop() public stretch?: BreakpointCustomizable<boolean> = false;
 
   /** Adapts the button color depending on the theme. */
-  @Prop() public theme?: Theme = 'light';
+  @Prop() public theme?: ThemeExtendedElectric = 'light';
 
   /** Add ARIA attributes. */
   @Prop() public aria?: SelectedAriaAttributes<ButtonAriaAttributes>;
@@ -120,6 +121,7 @@ export class ButtonPure {
       ['root']: true,
       ['root--loading']: this.loading && hasIcon,
       ['root--theme-dark']: isDark(this.theme),
+      ['root--theme-light-electric']: isLightElectric(this.theme),
       ['root--active']: this.active,
       ['root--with-icon']: hasIcon,
       ...mapBreakpointPropToClasses('root--size', this.size),