+++ packages/components/src/components/content/accordion/accordion.tsx
@@ -1,6 +1,6 @@
 import { Component, Element, Event, EventEmitter, Prop, Watch, h } from '@stencil/core';
-import { getPrefixedTagNames, isDark, mapBreakpointPropToClasses } from '../../../utils';
-import type { BreakpointCustomizable, Theme } from '../../../types';
+import { getPrefixedTagNames, isDark, isLightElectric, mapBreakpointPropToClasses } from '../../../utils';
+import type { BreakpointCustomizable, ThemeExtendedElectric } from '../../../types';
 import type { HeadlineTag } from '../../basic/typography/headline/headline-utils';
 import type { AccordionChangeEvent, AccordionSize } from './accordion-utils';
 import {
@@ -26,7 +26,7 @@ export class Accordion {
   @Prop() public size?: BreakpointCustomizable<AccordionSize> = 'small';
 
   /** Adapts the color when used on dark background. */
-  @Prop() public theme?: Theme = 'light';
+  @Prop() public theme?: ThemeExtendedElectric = 'light';
 
   /** Defines the heading used in accordion. */
   @Prop() public heading?: string;
@@ -103,6 +103,7 @@ export class Accordion {
     const rootClasses = {
       ['root']: true,
       ['root--theme-dark']: isDark(this.theme),
+      ['root--theme-light-electric']: isLightElectric(this.theme),
       ['root--open']: this.open,
       ...(!this.compact && mapBreakpointPropToClasses('root--size', this.size)),
       ['root--compact']: this.compact,